+++ CHANGELOG.md
@@ -11,6 +11,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - find way to create keyboard input for cli projects testing
 - find way to mock Scanner class
 
+## 2019-10-10
+### Changed
+- code cleanup and improvements in controller  and view in spring-gui-jdbc
+
+### Added
+- new tests for view in spring-gui-jdbc
+
+
 ## 2019-10-07
 ### Added
 - testing for spring-cli-jdbc controller